+++ changelog.md
@@ -12,6 +12,7 @@ All notable changes to Lock will be documented in this file. This file follows t
 - `.gitattributes` file
 - Added section to readme about testing your driver
 - The `Manager` class now has methods to set the lock instance for objects which implement the `LockAware` trait
+- `Lock` instances can now call a list of resource ids to which they are allowed or denied to perform the given action on
 
 ### Changed